refactor: prep technician base status readonly seam
